improve methods
