+++ models/contacts.json
@@ -70,5 +70,11 @@
     "name": "jakson Sweeny",
     "email": "jessye@gmail.com",
     "phone": "22222222222"
+  },
+  {
+    "id": "b92f55c2-3ac9-480f-8c67-0a068bf8597b",
+    "name": "Swany Swany",
+    "email": "swany@gmail.com",
+    "phone": "4938274239847"
   }
 ]
